+++ src/cmake.ts
@@ -9,6 +9,7 @@ import * as ajv from 'ajv';
 import * as vscode from 'vscode';
 
 import * as async from './async';
+import * as environment from './environment';
 import {ctest} from './ctest';
 import {FileDiagnostic,
         DiagnosticParser,
@@ -89,7 +90,7 @@ if(NOT is_set_up)
         _cmt_generate_system_info()
     endmacro()
 
-    if(NOT DEFINED CMAKE_BUILD_TYPE AND DEFINED CMAKE_CONFIGURATION_TYPES)
+    if({{{IS_MULTICONF}}})
         set(condition CONDITION "$<CONFIG:Debug>")
     endif()
 
@@ -432,6 +433,14 @@ export class ConfigurationReader {
     get testEnvironment(): Object {
         return this._readPrefixed<Object>('testEnvironment') || {};
     }
+
+    get defaultVariants(): Object {
+        return this._readPrefixed<Object>('defaultVariants') || {};
+    }
+
+    get ctestArgs(): string[] {
+        return this._readPrefixed<string[]>('ctestArgs') || [];
+    }
 }
 
 /**
@@ -616,6 +625,7 @@ export class CMakeTools {
     private _debugTargetButton = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 3.1);
     private _testStatusButton = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 3.05);
     private _warningMessage = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 3);
+    private _environmentSelectionButton = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 200);
     private _failingTestDecorationType = vscode.window.createTextEditorDecorationType({
         borderColor: 'rgba(255, 0, 0, 0.2)',
         borderWidth: '1px',
@@ -633,7 +643,7 @@ export class CMakeTools {
     });
     private _lastConfigureSettings = {};
     private _needsReconfigure = false;
-    private _workspaceCacheContent: util.WorkspaceCache;
+    private _workspaceCacheContent: util.WorkspaceCache = {};
     private _workspaceCachePath = path.join(vscode.workspace.rootPath || '~', '.vscode', '.cmaketools.json');
     private _targets: string[] = [];
     private _variantWatcher: vscode.FileSystemWatcher;
@@ -952,7 +962,15 @@ export class CMakeTools {
         if (!this._metaWatcher) {
             this._setupMetaWatcher();
         }
+        this._reloadVariants();
         this._refreshStatusBarItems();
+        this.testHaveCommand(this.config.cmakePath).then(exists => {
+            if (!exists) {
+                vscode.window.showErrorMessage(
+                    `Bad CMake executable "${this.config.cmakePath}". Is it installed and a valid executable?`
+                );
+            }
+        });
     }
 
     private _wsCacheWatcher: vscode.FileSystemWatcher;
@@ -1030,7 +1048,7 @@ export class CMakeTools {
                 variants = yaml.load(content);
             } catch(e) {
                 vscode.window.showErrorMessage(`${yaml_file} is syntactically invalid.`);
-                variants = util.DEFAULT_VARIANTS;
+                variants = this.config.defaultVariants;
             }
         } else if (await async.exists(json_file)) {
             const content = (await async.readFile(json_file)).toString();
@@ -1038,17 +1056,17 @@ export class CMakeTools {
                 variants = JSON.parse(content);
             } catch(e) {
                 vscode.window.showErrorMessage(`${json_file} is syntactically invalid.`);
-                variants = util.DEFAULT_VARIANTS;
+                variants = this.config.defaultVariants;
             }
         } else {
-            variants = util.DEFAULT_VARIANTS;
+            variants = this.config.defaultVariants;
         }
         const validated = validate(variants);
         if (!validated) {
             const errors = validate.errors as ajv.ErrorObject[];
             const error_strings = errors.map(err => `${err.dataPath}: ${err.message}`);
             vscode.window.showErrorMessage(`Invalid cmake-variants: ${error_strings.join('; ')}`);
-            variants = util.DEFAULT_VARIANTS;
+            variants = this.config.defaultVariants;
         }
         const sets = new Map() as util.VariantSet;
         for (const key in variants) {
@@ -1077,8 +1095,9 @@ export class CMakeTools {
         return this._buildVariants;
     }
     public set buildVariants(v : util.VariantSet) {
+        const before = this.activeVariant;
         this._buildVariants = v;
-        this._needsReconfigure = true;
+        this._needsReconfigure = JSON.stringify(this.activeVariant) !== JSON.stringify(before);
         this._refreshStatusBarItems();
     }
 
@@ -1128,7 +1147,6 @@ export class CMakeTools {
     }
     public set activeVariantCombination(v : util.VariantCombination) {
         this._activeVariantCombination = v;
-        this._needsReconfigure = true;
         this._workspaceCacheContent.variant = v;
         this._writeWorkspaceCacheContent().catch(CMakeTools._printUnhandledException);
         this._refreshStatusBarItems();
@@ -1198,11 +1216,99 @@ export class CMakeTools {
         });
     }
 
+    public activeEnvironments : string[] = [];
+    public activateEnvironment(name: string) {
+        const env = this.availableEnvironments.get(name);
+        if (!env) {
+            throw new Error(`Invalid environment named ${name}`);
+        }
+        if (!this.activeEnvironments) {
+            throw new Error(`Invalid state: Environments not yet loaded!`);
+        }
+        for (const other of this.availableEnvironments.values()) {
+            if (other.mutex === env.mutex && env.mutex !== undefined) {
+                const other_idx = this.activeEnvironments.indexOf(other.name);
+                if (other_idx >= 0) {
+                    this.activeEnvironments.splice(other_idx, 1);
+                }
+            }
+        }
+        this.activeEnvironments.push(name);
+        this._refreshStatusBarItems();
+        this._workspaceCacheContent.activeEnvironments = this.activeEnvironments;
+        this._writeWorkspaceCacheContent();
+    }
+
+    public deactivateEnvironment(name: string) {
+        if (!this.activeEnvironments) {
+            throw new Error('Invalid state: Environments not yet loaded!');
+        }
+        const idx = this.activeEnvironments.indexOf(name);
+        if (idx >= 0) {
+            this.activeEnvironments.splice(idx, 1);
+            this._refreshStatusBarItems();
+        }
+        this._workspaceCacheContent.activeEnvironments = this.activeEnvironments;
+        this._writeWorkspaceCacheContent();
+    }
+
+    private _availableEnvironments : Map<string, environment.Environment> = new Map();
+    public get availableEnvironments() : Map<string, environment.Environment> {
+        return this._availableEnvironments;
+    }
+
+    public async selectEnvironments(): Promise<void> {
+        const entries = Array.from(this.availableEnvironments.keys())
+            .map(name => ({
+                name: name,
+                label: this.activeEnvironments.indexOf(name) >= 0
+                    ? `$(check) ${name}`
+                    : name,
+                description: '',
+            }));
+        const chosen = await vscode.window.showQuickPick(entries);
+        if (!chosen) {
+            return;
+        }
+        this.activeEnvironments.indexOf(chosen.name) >= 0
+            ? this.deactivateEnvironment(chosen.name)
+            : this.activateEnvironment(chosen.name);
+    }
+
     private async _init(ctx: vscode.ExtensionContext): Promise<void> {
         this._channel = new ThrottledOutputChannel('CMake/Build');
         //this._channel = vscode.window.createOutputChannel('CMake/Build');
         this._ctestChannel = vscode.window.createOutputChannel('CTest Results');
         this._diagnostics = vscode.languages.createDiagnosticCollection('cmake-build-diags');
+
+        await this._refreshWorkspaceCacheContent();
+
+        // Start loading up available environments early, this may take a few seconds
+        const env_promises = environment.availableEnvironments();
+        for (const pr of env_promises) {
+            pr.then(env => {
+                if (env.variables) {
+                    console.log(`Detected available environemt "${env.name}"`);
+                    this._availableEnvironments.set(env.name, {
+                        name: env.name,
+                        variables: env.variables,
+                        mutex: env.mutex
+                    });
+                }
+            }).catch(e => {
+                debugger;
+                console.log('Error detecting environment', e);
+            });
+        }
+        await Promise.all(env_promises);
+        // All environments have been detected, now we can update the UI
+        this.activeEnvironments = [];
+        const envs = this._workspaceCacheContent.activeEnvironments || [];
+        envs.map(e => {
+            if (this.availableEnvironments.has(e)) {
+                this.activateEnvironment(e);
+            }
+        });
 
         const version_out = (await this.execute(['--version'], {
             silent: true,
@@ -1320,6 +1426,7 @@ export class CMakeTools {
                     this._debugButton.text = '$(bug)';
                     this._debugTargetButton.hide();
                 }
+                this._environmentSelectionButton.show();
             } else {
                 this._cmakeToolsStatusItem.hide();
                 this._buildButton.hide();
@@ -1327,6 +1434,7 @@ export class CMakeTools {
                 this._testStatusButton.hide();
                 this._debugButton.hide();
                 this._debugTargetButton.hide();
+                this._environmentSelectionButton.hide();
             }
             if (this._testStatusButton.text == '') {
                 this._testStatusButton.hide();
@@ -1365,6 +1473,22 @@ export class CMakeTools {
         this._debugButton.tooltip = 'Run the debugger on the selected target executable';
         this._debugTargetButton.text = this.currentDebugTarget || '[No target selected for debugging]';
         this._debugTargetButton.command = 'cmake.selectDebugTarget';
+        this._environmentSelectionButton.command = 'cmake.selectEnvironments';
+
+        if (this.activeEnvironments !== null) {
+            if (this.activeEnvironments.length) {
+                this._environmentSelectionButton.text = `Working in ${this.activeEnvironments.join(', ')}`;
+            } else {
+                if (this.availableEnvironments.size !== 0) {
+                    this._environmentSelectionButton.text = 'Select a build environment...';
+                } else {
+                    // No environments available. No need to show this button
+                    this._environmentSelectionButton.hide();
+                }
+            }
+        } else {
+            this._environmentSelectionButton.text = 'Detecting available build environments...';
+        }
     }
 
     public get projectName() {
@@ -1505,9 +1629,9 @@ export class CMakeTools {
     /**
      * @brief Determine if the project is using a multi-config generator
      */
-    public get isMultiConf() {
+    public get isMultiConf(): boolean {
         const gen = this.activeGenerator;
-        return gen && util.isMultiConfGenerator(gen);
+        return !!gen && util.isMultiConfGenerator(gen);
     }
 
     public get activeGenerator(): Maybe<string> {
@@ -1543,19 +1667,32 @@ export class CMakeTools {
                     },
                    parser: util.OutputParser = new NullParser())
     : Promise<ExecutionResult> {
-        return new Promise<ExecutionResult>((resolve, _) => {
+        return new Promise<ExecutionResult>(async (resolve) => {
             const silent: boolean = options && options.silent || false;
             console.info('Execute cmake with arguments:', args);
             let stdout = '';
             let stderr = '';
             const pipe = proc.spawn(this.config.cmakePath, args, {
                 env: Object.assign(
-                    Object.assign({
+                    {
                         // We set NINJA_STATUS to force Ninja to use the format
                         // that we would like to parse
                         NINJA_STATUS: '[%f/%t %p] '
-                    }, options.environment),
+                    },
+                    options.environment,
                     this.config.environment,
+                    this.activeEnvironments.reduce<any>(
+                        (acc, name) => {
+                            const env_ = this.availableEnvironments.get(name);
+                            console.assert(env_);
+                            const env = env_!;
+                            for (const entry of env.variables.entries()) {
+                                acc[entry[0]] = entry[1];
+                            }
+                            return acc;
+                        },
+                        {}
+                    ),
                     process.env
                 )
             });
@@ -1703,15 +1840,29 @@ export class CMakeTools {
         return null;
     }
 
-    private async _prebuild() {
+    private async _prebuild(): Promise<boolean> {
         if (this.config.clearOutputBeforeBuild) {
             this._channel.clear();
         }
 
         if (this.config.saveBeforeBuild && vscode.workspace.textDocuments.some(doc => doc.isDirty)) {
             this._channel.appendLine("[vscode] Saving unsaved text documents...");
-            await vscode.workspace.saveAll();
+            const is_good = await vscode.workspace.saveAll();
+            if (!is_good) {
+                const chosen = await vscode.window.showErrorMessage<vscode.MessageItem>(
+                    'Not all open documents were saved. Would you like to build anyway?',
+                    {
+                        title: 'Yes',
+                        isCloseAffordance: false,
+                    },
+                    {
+                        title: 'No',
+                        isCloseAffordance: true,
+                    });
+                return chosen.title === 'Yes';
+            }
         }
+        return true;
     }
 
     public get numJobs(): number {
@@ -1756,8 +1907,12 @@ export class CMakeTools {
             }
         }
 
-        if (run_prebuild)
-            await this._prebuild();
+        if (run_prebuild) {
+            const ok = await this._prebuild();
+            if (!ok) {
+                return -1;
+            }
+        }
 
         const cmake_cache = this.cachePath;
         this._channel.show();
@@ -1802,19 +1957,25 @@ export class CMakeTools {
         }
 
         if (!(await async.exists(this.binaryDir))) {
-            await fs.mkdir(this.binaryDir);
+            await util.ensureDirectory(this.binaryDir);
         }
 
         const cmt_dir = path.join(this.binaryDir, 'CMakeTools');
         if (!(await async.exists(cmt_dir))) {
-            await fs.mkdir(cmt_dir);
+            await util.ensureDirectory(cmt_dir);
         }
 
         if (!this.serverClient) {
             // We only need to generate our own metadata if we do not have
             // CMake Server support
             const helpers = path.join(cmt_dir, 'CMakeToolsHelpers.cmake');
-            await util.writeFile(helpers, CMAKETOOLS_HELPER_SCRIPT);
+            const helper_content = util.replaceAll(CMAKETOOLS_HELPER_SCRIPT,
+                                            '{{{IS_MULTICONF}}}',
+                                            is_multi_conf
+                                                ? '1'
+                                                : '0'
+                                            );
+            await util.writeFile(helpers, helper_content);
             const old_path = settings['CMAKE_PREFIX_PATH'] as Array<string> || [];
             settings['CMAKE_MODULE_PATH'] = Array.from(old_path).concat([
                 cmt_dir.replace(/\\/g, path.posix.sep)
@@ -1907,6 +2068,7 @@ export class CMakeTools {
         if (!result.retc) {
             await this._refreshAll();
             await this._reloadConfiguration();
+            this._needsReconfigure = false;
         }
         return result.retc;
     }
@@ -1941,7 +2103,10 @@ export class CMakeTools {
         if (!target) {
             throw new Error('Unable to determine target to build. Something has gone horribly wrong!');
         }
-        await this._prebuild();
+        const ok = await this._prebuild();
+        if (!ok) {
+            return -1;
+        }
         if (this._needsReconfigure) {
             const retc = await this.configure([], false);
             if (!!retc)
@@ -2091,8 +2256,12 @@ export class CMakeTools {
     }
 
     public async setBuildType(): Promise<Number> {
-        await this.setBuildTypeWithoutConfigure();
-        return await this.configure();
+        const do_configure = await this.setBuildTypeWithoutConfigure();
+        if (do_configure) {
+            return await this.configure();
+        } else {
+            return -1;
+        }
     }
 
     public async debugTarget() {
@@ -2160,8 +2329,8 @@ export class CMakeTools {
                     'ctest', '-j' + this.numCTestJobs,
                     '-C', this.selectedBuildType || 'Debug',
                     '-T', 'test',
-                    '--output-on-failure'
-                ],
+                    '--output-on-failure',
+                ].concat(this.config.ctestArgs),
                 {
                     silent: false,
                     environment: this.config.testEnvironment,